Implements: search filter in ArticleList, search button action. Changes: ArticleCard link to SingleArticlePage implementation.
SearchBar and ArticleList: Added boolean prop to ArticleList, which is used as condition in render method. If condition is true, articleList collection is filtered using Array.filter(item => item.content.includes(searchString).
This makes for a very primitive search, based solely on exact, case sensitive matching of the search box input string.
The magnifying glass icon uses the Router.useHistory().push() hook to trigger routing to "/search/:search_string". This lets the user return to /home to unfilter the article list, and also keeps a history of search inputs using the browser history.
The search bar input element triggers the onClick function programmatically on hitting "Enter".
Finally, the New Article FAB has been attached to ArticleList, instead of the Routes in App.

ArticleCard, and link to SingleArticlePage: The ArticleCard now uses the history.push() hook, instead of wrapping the card component in a <Link>.