Implements article editing.
Articles can now be edited, accessible through the FAB on a single article page.
Editing is implemented through the NewArticleForm, by routing to "/editPost/:postId" and passing react-router 'match' as prop to the comp. the NewArticleForm comp now test for the presence of 'match.props', which is only defined if 'match' was passed in the render part of Router.route. The artId of the article to be edited is extracted as 'match.props.params.artId'.
NewArticleForm either starts blank, or loads the title and content of the artId specified in the 'match' prop.
The NewArticleForm now also shows a 'Delete Post' button, only if the content of an existing article has been loaded. This is implemented through the tailwind '.hidden' class, and should probably be reconsidered...